refactor(orch): move memfile dedup into its own file
Extract the dedup classification, fetch-window promotion, and drain logic out
of the generic block cache into dedup.go (same package, no behavior change),
leaving Cache.Dedup as the hook so the strategy is contained and easier to
swap or compare later.
